refactor: Code style and import paths
