+++ src/components/HomeView/index.ts
@@ -1,0 +1,3 @@
+export { default as HomeViewProductGrid } from './HomeViewProductGrid/HomeViewProductGrid.vue'
+export { default as HomeViewProductCard } from './HomeViewProductCard/HomeViewProductCard.vue'
+export { default as HomeViewPagination } from './HomeViewPagination/HomeViewPagination.vue'
