+++ src/components/ui/index.ts
@@ -1,0 +1,2 @@
+export { default as UiButton } from './Button/UiButton.vue'
+export { default as UiPopup } from './Popup/UiPopup.vue'
